temp_files/requirements.txt
aiofiles
annotated-types
anyio
asttokens
attrs
blinker
certifi
charset-normalizer
click
colorama
comm
ConfigArgParse
contourpy
cycler
dash
decorator
executing
fastapi
fastjsonschema
ffmpy
filelock
Flask
fonttools
fsspec
gdown
gradio
gradio_client
gradio_imageslider
groovy
h11
httpcore
httpx
huggingface-hub
idna
importlib_metadata
itsdangerous
jedi
Jinja2
jsonschema
jsonschema-specifications
jupyter_core
jupyterlab_widgets
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mpmath
narwhals
nbformat
nest-asyncio
networkx
numpy<2
open3d
opencv-python
orjson
packaging
pandas
parso
pillow
platformdirs
plotly
prompt_toolkit
pure_eval
pydantic_core
pydub
Pygments
pyparsing
python-dateutil
python-multipart
pytz
PyYAML
referencing
requests
retrying
rich
rpds-py
ruff
safehttpx
scikit-image
semantic-version
setuptools
shellingham
six
sniffio
stack-data
starlette
sympy
tensorflow<2.11
tensorflow_hub
tomlkit
torch
torchvision
tqdm
traitlets
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
wcwidth
websockets
Werkzeug
wheel
widgetsnbextension
zipp
pydantic==2.10.6